fix bug with empty photo

+++ src/Flyers/Croco/Photo.php
@@ -18,6 +18,9 @@ class Photo
         curl_setopt($ch, CURLOPT_FOLLOWLOCATION, 1);
         $raw = curl_exec($ch);
         curl_close ($ch);
+        if (strlen($raw) == 0) {
+            return $this->grab($url, $name);
+        }
         if (file_exists($name)) unlink($name);
         $fp = fopen($name,'x');
         fwrite($fp, $raw);
